Move array_from_* and mask_from_* methods and tests
